bootloader: archive: switch to 64-bit integers for size-related fields
Similar to previous switch from signed 32-bit to unsigned 32-bit
integers, we now switch signed 64-bit integers. This allows
5 GB variant of test_large_data_file to pass on 64-bit linux.

+++ bootloader/src/pyi_pythonlib.c
@@ -615,7 +615,7 @@ int
 pyi_pylib_install_zlib(ARCHIVE_STATUS *status, TOC *ptoc)
 {
     int rc = 0;
-    unsigned int zlibpos = status->pkgstart + ptoc->pos;
+    int64_t zlibpos = status->pkgstart + ptoc->pos;
     PyObject * sys_path, *zlib_entry, *archivename_obj;
 
     /* Note that sys.path contains PyUnicode on py3. Ensure
@@ -633,7 +633,7 @@ pyi_pylib_install_zlib(ARCHIVE_STATUS *status, TOC *ptoc)
      */
     archivename_obj = PI_PyUnicode_DecodeFSDefault(status->archivename);
 #endif
-    zlib_entry = PI_PyUnicode_FromFormat("%U?%u", archivename_obj, zlibpos);
+    zlib_entry = PI_PyUnicode_FromFormat("%U?%lli", archivename_obj, zlibpos);
     PI_Py_DecRef(archivename_obj);
 
     sys_path = PI_PySys_GetObject("path");
